Handle better the delayed removal of a treeiter
When a stream is detached, it may happen that it is remapped to another circuit
before that it is removed from out treestore before the timeout set for its
actual deletion happen.

To prevent removing an invalid iter, we now add a remove_stream_delayed method
that gets the treeiter corresponding to the stream, plans its deletion and
remove it from the known iters. We then test wether the treeiter isn't valid
before the actual removal happen.

For coherence, we do the same for circuits.