Stop Git-ignoring the POT file.
The way we use Transifex, we need to point Tor people to a POT file in
Git that they can regularly pull. I see that Onion Circuits has no such
file in Git. It's a bit sad to store such auto-generated files in Git,
but it's even more sad to ship software without translations.

refs: Tails#11173